+++ app/views/decidim/time_tracker/admin/tasks/index.html.erb
@@ -9,7 +9,7 @@
     </h2>
   </div>
 
-    <div class="card-section">
+  <div class="card-section">
     <div class="table-scroll">
       <table class="table-list">
         <thead>
@@ -45,16 +45,57 @@
                 <%= task.assignees_count(filter: :pending) %><br>
               </td>
               <td class="table-list__actions">
-                <% if allowed_to? :update, :questionnaire, questionnaire: task.questionnaire %>
-                  <%= icon_link_to "document", edit_task_form_path(task), t("tasks.actions.questionnaire", scope: "decidim.time_tracker.admin"), class: "action-icon--edit#{' action-icon-success' if task.has_questions?}" %>
-                <% end %>
-
                 <% if allowed_to? :update, :task, task: task %>
                   <%= icon_link_to "pencil", edit_task_path(task), t("tasks.actions.edit", scope: "decidim.time_tracker.admin"), class: "action-icon--edit" %>
                 <% end %>
 
                 <% if allowed_to? :destroy, :task, task: task %>
                   <%= icon_link_to "circle-x", task_path(task), t("tasks.actions.destroy", scope: "decidim.time_tracker.admin"), method: :delete, class: "action-icon--remove", data: { confirm: t("tasks.actions.confirm_destroy", scope: "decidim.time_tracker.admin") } %>
+                <% end %>
+              </td>
+            </tr>
+          <% end %>
+        </tbody>
+      </table>
+    </div>
+  </div>
+</div>
+
+<div class="card">
+  <div class="card-divider">
+    <h2 class="card-title">
+      <%= t(".questionnaires") %>
+      <div class="button--title">
+        <% if allowed_to? :update, :questionnaire, questionnaire: time_tracker.questionnaire %>
+          <%= link_to t("tasks.actions.questionnaire", scope: "decidim.time_tracker.admin"), edit_questionnaire_path, class: "button tiny button--simple" %>
+        <% end %>
+      </div>
+    </h2>
+  </div>
+
+  <div class="card-section">
+    <div class="table-scroll">
+      <table class="table-list">
+        <thead>
+          <tr>
+            <th><%= t("models.questionnaire.fields.name", scope: "decidim.time_tracker") %></th>
+            <th><%= t("models.questionnaire.fields.title", scope: "decidim.time_tracker") %></th>
+            <th class="actions"><%= t("actions.title", scope: "decidim.time_tracker.admin") %></th>
+          </tr>
+        </thead>
+        <tbody>
+          <% [:questionnaire].each do |questionnaire_name| %>
+            <% questionnaire = time_tracker.send(questionnaire_name) %>
+            <tr data-id="<%= questionnaire.id %>">
+              <td>
+                <%= t(".#{questionnaire_name}") %><br>
+              </td>
+              <td>
+                <%= translated_attribute(questionnaire.title) %><br>
+              </td>
+              <td class="table-list__actions">
+                <% if allowed_to? :update, :questionnaire, questionnaire: questionnaire %>
+                  <%= icon_link_to "pencil", edit_questionnaire_path, t("tasks.actions.edit", scope: "decidim.time_tracker.admin"), class: "action-icon--edit" %>
                 <% end %>
               </td>
             </tr>
